% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{perspectiveOutput}
\alias{perspectiveOutput}
\title{Shiny Output for Perspective Viewer}
\usage{
perspectiveOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable name.}

\item{width}{CSS width (default \code{"100\%"}).}

\item{height}{CSS height (default \code{"400px"}).}
}
\value{
A Shiny output element.
}
\description{
Creates a Perspective viewer output element for use in a Shiny UI.
}
\examples{
\dontrun{
library(shiny)
ui <- fluidPage(
  perspectiveOutput("viewer", height = "600px")
)
}

}
