% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{perspectiveProxy}
\alias{perspectiveProxy}
\title{Create a Perspective Proxy Object for Shiny}
\usage{
perspectiveProxy(session, outputId)
}
\arguments{
\item{session}{The Shiny session object (usually \code{session}).}

\item{outputId}{The output ID of the Perspective widget to control.}
}
\value{
A proxy object of class \code{"perspective_proxy"}.
}
\description{
Creates a proxy object that can be used to update an existing Perspective
viewer in a Shiny app without re-rendering the entire widget. Use this with
\code{psp_update}, \code{psp_replace}, \code{psp_clear}, \code{psp_restore},
and \code{psp_reset} to modify the viewer.
}
\examples{
if (interactive()) {
server <- function(input, output, session) {
  output$viewer <- renderPerspective({
    perspective(mtcars)
  })

  observeEvent(input$add_data, {
    proxy <- perspectiveProxy(session, "viewer")
    psp_update(proxy, new_data)
  })
}
}

}
