% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_clear}
\alias{psp_clear}
\title{Clear All Data from a Perspective Viewer}
\usage{
psp_clear(proxy)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Removes all rows from the Perspective table (schema is preserved).
}
