% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_columns}
\alias{psp_columns}
\title{Get Table Column Names}
\usage{
psp_columns(proxy)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Requests the column names of the Perspective table. The result is
delivered asynchronously to \code{input$<outputId>_columns}.
}
\examples{
\dontrun{
proxy <- perspectiveProxy(session, "viewer")
psp_columns(proxy)
}

}
