% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_remove}
\alias{psp_remove}
\title{Remove Rows by Key from a Perspective Viewer}
\usage{
psp_remove(proxy, keys)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}

\item{keys}{A vector of key values identifying the rows to remove.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Removes rows matching the given primary-key values from an indexed
Perspective table. The table must have been created with an \code{index}
column (see \code{\link{perspective}}).
}
