% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_replace}
\alias{psp_replace}
\title{Replace All Data in a Perspective Viewer}
\usage{
psp_replace(proxy, data, use_arrow = FALSE)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}

\item{data}{A data.frame containing the replacement data.}

\item{use_arrow}{Logical; use Arrow IPC serialization. Default \code{FALSE}.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Replaces the entire dataset in the Perspective table.
}
\examples{
\dontrun{
proxy <- perspectiveProxy(session, "viewer")
psp_replace(proxy, updated_data)
}

}
