% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_reset}
\alias{psp_reset}
\title{Reset Viewer to Default State}
\usage{
psp_reset(proxy)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Resets the Perspective viewer to its default configuration.
}
\examples{
if (interactive()) {
proxy <- perspectiveProxy(session, "viewer")
psp_reset(proxy)
}

}
