% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_restore}
\alias{psp_restore}
\title{Restore Viewer Configuration}
\usage{
psp_restore(proxy, config)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}

\item{config}{A list of Perspective viewer configuration options.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Applies a configuration object to the Perspective viewer, changing columns,
group_by, split_by, filters, sort, aggregates, plugin, etc.
}
\examples{
\dontrun{
proxy <- perspectiveProxy(session, "viewer")
psp_restore(proxy, list(plugin = "Y Bar", group_by = list("cyl")))
}

}
