% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_size}
\alias{psp_size}
\title{Get Table Row Count}
\usage{
psp_size(proxy)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Requests the number of rows in the Perspective table. The result is
delivered asynchronously to \code{input$<outputId>_size}.
}
\examples{
if (interactive()) {
proxy <- perspectiveProxy(session, "viewer")
psp_size(proxy)
}

}
