% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{psp_update}
\alias{psp_update}
\title{Update (Append) Data in a Perspective Viewer}
\usage{
psp_update(proxy, data, use_arrow = FALSE)
}
\arguments{
\item{proxy}{A \code{\link{perspectiveProxy}} object.}

\item{data}{A data.frame of new rows to append.}

\item{use_arrow}{Logical; use Arrow IPC serialization. Default \code{FALSE}.}
}
\value{
The proxy object (invisibly), for chaining.
}
\description{
Sends new rows to be appended to the existing Perspective table.
}
\examples{
if (interactive()) {
proxy <- perspectiveProxy(session, "viewer")
psp_update(proxy, new_data)
}

}
