% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perspective-shiny.R
\name{renderPerspective}
\alias{renderPerspective}
\title{Render a Perspective Viewer in Shiny}
\usage{
renderPerspective(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a \code{\link{perspective}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Logical; is \code{expr} a quoted expression?}
}
\value{
A Shiny render function.
}
\description{
Server-side rendering function for the Perspective widget.
}
\examples{
\dontrun{
server <- function(input, output) {
  output$viewer <- renderPerspective({
    perspective(mtcars, group_by = "cyl", plugin = "Y Bar")
  })
}
}

}
